requirements.txt
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pypdfium2==4.30.0
pyTelegramBotAPI==4.18.1
requests==2.32.2
telebot==0.0.5
urllib3==2.2.1
Werkzeug==3.0.3
fpdf==1.7.2
python-dotenv==1.0.1